+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Added
 * StackPanel, a control which behaves like the original stack panel but with the additional property for a spacing between the children
 * UniformPanel, A UniformGrid with only one row or one column, depending on the orientation, which adds a spacing between the items
+* ItemsPanel, a combined control between StackPanel and UniformPanel
 
 ## [1.6.0] - 2021-07-25
 ### Added